network: Bridge - Add support for a DHCP Relay Agent
A DHCP relay daemon will be started that will forward all DHCP/BOOTP
requests on the bridge network via the first declared forward
interface. The relay is broadcast rather than routed so no IP address
is needed on the bridge.

The XML <relay> element's "relay" property of the active DHCP stanza
defaults to 'no'. Set it to 'yes' to enable the relay:

<ip ...>
 <dhcp relay='yes'/>
</ip>

The relay will not be started if the "enable" property is 'no':

<ip ...>
 <dhcp enable='no' relay='yes'/>
</ip>

Signed-off-by: TJ <linux@iam.tj>